split methods in fork to separate file
